[AMXD-16] Cadent Bid Adapter: Rename files
